+++ src/main/java/settleup/backend/domain/transaction/entity/OptimizedTransactionEntity.java
@@ -6,6 +6,7 @@ import settleup.backend.domain.group.entity.GroupEntity;
 import settleup.backend.domain.user.entity.UserEntity;
 import settleup.backend.global.common.Status;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 
@@ -34,7 +35,7 @@ public class OptimizedTransactionEntity implements TransactionalEntity {
     private UserEntity recipientUser;
 
     @Column(name = "transaction_amount", nullable = false)
-    private double transactionAmount;
+    private BigDecimal transactionAmount;
 
     @Column(name = "optimization_status",nullable = false)
     @Enumerated(EnumType.STRING)
@@ -102,6 +103,6 @@ public class OptimizedTransactionEntity implements TransactionalEntity {
     }
 
     @Override
-    public double getTransactionAmount(){return  this.transactionAmount;}
+    public BigDecimal getTransactionAmount(){return  this.transactionAmount;}
 
 }
