+++ src/main/java/settleup/backend/domain/transaction/entity/UltimateOptimizedTransactionEntity.java
@@ -7,6 +7,7 @@ import settleup.backend.domain.group.entity.GroupEntity;
 import settleup.backend.domain.user.entity.UserEntity;
 import settleup.backend.global.common.Status;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -34,7 +35,7 @@ public class UltimateOptimizedTransactionEntity implements TransactionalEntity {
     private UserEntity recipientUser;
 
     @Column(name = "transaction_amount", nullable = false)
-    private double transactionAmount;
+    private BigDecimal transactionAmount;
 
     @Column(name = "optimization_status",nullable = false)
     @Enumerated(EnumType.STRING)
@@ -82,7 +83,7 @@ public class UltimateOptimizedTransactionEntity implements TransactionalEntity {
     }
 
     @Override
-    public double getTransactionAmount() {
+    public BigDecimal getTransactionAmount() {
         return this.transactionAmount;
     }
