+++ src/main/java/settleup/backend/domain/transaction/entity/dto/IntermediateCalcDto.java
@@ -8,6 +8,7 @@ import settleup.backend.domain.transaction.entity.RequiresTransactionEntity;
 import settleup.backend.domain.transaction.entity.TransactionalEntity;
 import settleup.backend.domain.user.entity.UserEntity;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -17,7 +18,7 @@ public class IntermediateCalcDto {
     private GroupEntity group;
     private UserEntity SenderUser;
     private UserEntity recipientUser;
-    private double transactionAmount;
+    private BigDecimal transactionAmount;
     private List<RequiresTransactionEntity> duringOptimizationUsed;
     private List<TransactionalEntity> duringFinalOptimizationUsed;
 }
